remove other components
